requirements.txt
pymongo
requests
psycopg2-binary
streamlit
pandas
plotly
speedtest-cli
beautifulsoup4
